third_party/blink/web_tests/fast/css-grid-layout/grid-template-areas-get-set-expected.txt

This test checks that grid-template-areas is properly parsed.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Test getting grid-template-areas set through CSS.
PASS window.getComputedStyle(gridWithDefaultTemplate).getPropertyValue('grid-template-areas') is "none"
PASS window.getComputedStyle(gridWithSingleStringTemplate).getPropertyValue('grid-template-areas') is "\"area\""
PASS window.getComputedStyle(gridWithTwoColumnsTemplate).getPropertyValue('grid-template-areas') is "\"first second\""
PASS window.getComputedStyle(gridWithTwoRowsTemplate).getPropertyValue('grid-template-areas') is "\"first\" \"second\""
PASS window.getComputedStyle(gridWithSpanningColumnsTemplate).getPropertyValue('grid-template-areas') is "\"span span\""
PASS window.getComputedStyle(gridWithSpanningRowsDotTemplate).getPropertyValue('grid-template-areas') is "\"span\" \".\""
PASS window.getComputedStyle(gridWithSpanningRowsDotsTemplate).getPropertyValue('grid-template-areas') is "\"span\" \".\""
PASS window.getComputedStyle(gridWithDotColumn).getPropertyValue('grid-template-areas') is "\"header .\" \"footer .\""
PASS window.getComputedStyle(gridWithDotsColumn).getPropertyValue('grid-template-areas') is "\"header .\" \"footer .\""
PASS window.getComputedStyle(gridWithDotCells).getPropertyValue('grid-template-areas') is "\"first .\" \". second\""
PASS window.getComputedStyle(gridWithDotsCells).getPropertyValue('grid-template-areas') is "\"first .\" \". second\""
PASS window.getComputedStyle(gridWithComplexDotAreas).getPropertyValue('grid-template-areas') is "\". header header .\" \". . . .\" \"nav main main aside\" \"nav main main aside\" \". . . aside\" \". footer footer aside\""
PASS window.getComputedStyle(gridWithComplexDotsAreas).getPropertyValue('grid-template-areas') is "\". header header .\" \". . . .\" \"nav main main aside\" \"nav main main aside\" \". . . aside\" \". footer footer aside\""
PASS window.getComputedStyle(gridWithMixedDotAndDotsAreas).getPropertyValue('grid-template-areas') is "\". title .\" \". main main\" \"nav . aside\" \". footer .\""
PASS window.getComputedStyle(gridWithDotsNoSpace).getPropertyValue('grid-template-areas') is "\". title . nav . . main test\""
PASS window.getComputedStyle(gridWithHorizontalRectangle).getPropertyValue('grid-template-areas') is "\"a a a\" \"a a a\""
PASS window.getComputedStyle(gridWithVerticalRectangle).getPropertyValue('grid-template-areas') is "\"a a\" \"a a\" \"a a\""
Test grid-template-areas: initial
PASS window.getComputedStyle(element).getPropertyValue('grid-template-areas') is "\"foobar\""
PASS window.getComputedStyle(element).getPropertyValue('grid-template-areas') is "none"
PASS window.getComputedStyle(element).getPropertyValue('grid-template-areas') is "\"foobar\""
PASS window.getComputedStyle(element).getPropertyValue('grid-template-areas') is "none"
Test grid-template-areas: inherit
PASS window.getComputedStyle(parentElement).getPropertyValue('grid-template-areas') is "\"foo bar\""
PASS window.getComputedStyle(parentElement).getPropertyValue('grid-template-areas') is "\"foo bar\""
PASS window.getComputedStyle(parentElement).getPropertyValue('grid-template-areas') is "\"foo bar\""
PASS window.getComputedStyle(element).getPropertyValue('grid-template-areas') is "\"foo bar\""
Test invalid grid-template-areas values.
PASS window.getComputedStyle(element).getPropertyValue('grid-template-areas') is "none"
PASS window.getComputedStyle(element).getPropertyValue('grid-template-areas') is "none"
PASS window.getComputedStyle(element).getPropertyValue('grid-template-areas') is "none"
PASS window.getComputedStyle(element).getPropertyValue('grid-template-areas') is "none"
PASS window.getComputedStyle(element).getPropertyValue('grid-template-areas') is "none"
PASS window.getComputedStyle(element).getPropertyValue('grid-template-areas') is "none"
PASS window.getComputedStyle(element).getPropertyValue('grid-template-areas') is "none"
PASS window.getComputedStyle(element).getPropertyValue('grid-template-areas') is "none"
PASS window.getComputedStyle(element).getPropertyValue('grid-template-areas') is "none"


FIXME: We currently don't validate that the named grid areas are <indent>.
FAIL window.getComputedStyle(element).getPropertyValue('grid-template-areas') should be none. Was "nav-up".
PASS successfullyParsed is true

TEST COMPLETE
